Add Terarium CRUD

+++ src/main/java/terarium/server/service/AnimalService.java
@@ -3,15 +3,18 @@ package terarium.server.service;
 import java.util.List;
 import java.util.Optional;
 
+import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import terarium.server.model.Animal;
-import terarium.server.dto.Animal.UpdateAnimalDto;
+import terarium.server.ServerApplication;
 import terarium.server.repository.AnimalRepository;
 
 @Service
 public class AnimalService {
+    private Logger log = ServerApplication.log;
+    
     @Autowired
     private AnimalRepository animalRepository;
     
@@ -31,7 +34,7 @@ public class AnimalService {
         try {
             return animalRepository.save(animal);
         } catch (Error e) {
-            System.err.println(e.getMessage());
+            log.error("Crete Animal error", e);
             
             return null;
         }
@@ -40,20 +43,18 @@ public class AnimalService {
     public Animal DeleteAnimal(int id){
         Optional<Animal> animal = animalRepository.findById(id);
         
+        if (animal.isEmpty()) return null;
+        
         animalRepository.deleteById(id);
         
-        return animal.isPresent() ? animal.get() : null;
+        return animal.get();
     }
     
-    public Animal UpdateAnimal(UpdateAnimalDto updateAnimalDto, int id){
+    public Animal UpdateAnimal(Animal animal){
         try {
-            Animal animal = new Animal().FromDto(updateAnimalDto);
-            
-            animal.setId(id);
-            
             return animalRepository.save(animal);
         } catch (Error e) {
-            System.err.println(e.getMessage());
+            log.error("Update Animal error", e);
             
             return null;
         }
